requirements.txt
streamlit>=1.35.0
anthropic>=0.39.0
requests>=2.31.0
fpdf2>=2.7.0
